Welease 20.
Update gnu method to take an “email” property rather than “notes”, so
the latter can be shorter, and the former can supply an entire email
message, including headers.

Cope with FreshMeat becoming Freecode.

Remove note saying that “notes” file need not exist, as it does need to.

Fix a bug in the lua method: add a subject so that the rockspec
name is used as an attachment name, not as the message body.