Wemove a FIXME
One of the extra releasers proposed was personal (this is now meant to be
usable by others), the other rather out of date (web sites that are file
based should probably be using git or similar now, not just rsync).

In any case, I’ve not found the will to add them in many years, so probably
never would’ve.